+++ src/Query/Constraint/PrimaryKey.php
@@ -1,0 +1,35 @@
+<?php
+declare(strict_types = 1);
+
+namespace Formal\AccessLayer\Query\Constraint;
+
+use Formal\AccessLayer\Table\Column;
+
+/**
+ * @psalm-immutable
+ */
+final class PrimaryKey
+{
+    private Column\Name $column;
+
+    private function __construct(Column\Name $column)
+    {
+        $this->column = $column;
+    }
+
+    /**
+     * @psalm-pure
+     */
+    public static function on(Column\Name $column): self
+    {
+        return new self($column);
+    }
+
+    /**
+     * @return non-empty-string
+     */
+    public function sql(): string
+    {
+        return "PRIMARY KEY ({$this->column->sql()})";
+    }
+}